Remove plog completely from decoder
Prior to this commit, there were still two functions that were
specialized to plog classes.  These classes are actually not that
special to plog, and this commit has moved them to polysync/decoder/types.hpp.  Other edits finish the job to get the tests to pass.
